printf: int <-> integer
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2185 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/image/pdftoepdf.cc
@@ -877,7 +877,7 @@ static void write_epdf1(image_dict * idict)
         if (lastGroup->dictGetLength() > 0) {
             pdf_puts("/Group ");
             groupIsIndirect = lastGroup->isRef();
-            pdf_printf("%d 0 R", epdf_lastGroupObjectNum);
+            pdf_printf("%d 0 R", (int)epdf_lastGroupObjectNum);
             pdf_puts("\n");
         }
       }
